add new lib

+++ rust/rust_lib/print_number.py
@@ -3,4 +3,4 @@ import ctypes
 rust_lib = ctypes.CDLL("target/release/librust_lib.so")
 
 if __name__ == "__main__":
-    rust_lib.print_number(ctypes.c_int(65234234))
+    rust_lib.print_number(65234234)
